thirdparty: support import brpc and gtest lib from cmake
Summary:

1. support import brpc lib by cmake
2. bump brpc version to 1.8.0
3. bump gtest version to 1.14.0
4. fix compile warnings

details: https://github.com/ehds/braft/pull/5

+++ src/braft/file_system_adaptor.cpp
@@ -122,19 +122,21 @@ ssize_t BufferedSequentialWriteFileAdaptor::write(const butil::IOBuf& data, off_
     BRAFT_VLOG << "begin write offset " << offset << ", data_size " << data.size()
               << ", buffer_offset " << _buffer_offset 
               << ", buffer_size " << _buffer_size;
-    if (offset < _buffer_offset + _buffer_size) {
-        LOG(WARNING) << "Fail to write into buffered file adaptor with invalid range"
-                     << ", offset: " << offset
-                     << ", data_size: " << data.size()
-                     << ", buffer_offset: " << _buffer_offset
-                     << ", buffer_size: " << _buffer_size;
-        errno = EINVAL;
-        return -1;
-    } else if (offset > _buffer_offset + _buffer_size) {
-        // passby hole
-        CHECK(_buffer_size == 0);
-        BRAFT_VLOG << "seek to new offset " << offset << " as there is hole";
-        seek(offset);
+    if (static_cast<size_t>(offset) <
+        static_cast<size_t>(_buffer_offset) + _buffer_size) {
+      LOG(WARNING)
+          << "Fail to write into buffered file adaptor with invalid range"
+          << ", offset: " << offset << ", data_size: " << data.size()
+          << ", buffer_offset: " << _buffer_offset
+          << ", buffer_size: " << _buffer_size;
+      errno = EINVAL;
+      return -1;
+    } else if (static_cast<size_t>(offset) >
+               static_cast<size_t>(_buffer_offset) + _buffer_size) {
+      // passby hole
+      CHECK(_buffer_size == 0);
+      BRAFT_VLOG << "seek to new offset " << offset << " as there is hole";
+      seek(offset);
     }
     const size_t saved_size = data.size();
     _buffer.append(data);
